requirements.txt
asgiref==3.5.0
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
greenlet==1.1.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.7.1
parsel==1.6.0
priority==1.3.0
Protego==0.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
python-scrapyd-api==2.1.2
pytz==2021.3
queuelib==1.6.2
requests==2.27.1
Scrapy==2.5.1
scrapyd==1.3.0
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
Twisted==22.1.0
typing_extensions==4.1.1
urllib3==1.26.8
w3lib==1.22.0
zope.interface==5.4.0
